人脸id关联

+++ GeoPic/src/semester/cn/servlets/getPhotoDataServlet.java
@@ -1,5 +1,6 @@
 package semester.cn.servlets;
 
+import net.sf.json.JSONObject;
 import org.opencv.core.*;
 import org.opencv.imgcodecs.Imgcodecs;
 import org.opencv.imgproc.Imgproc;
@@ -29,61 +30,57 @@ import static org.opencv.imgcodecs.Imgcodecs.imwrite;
 
 @WebServlet(name = "getPhotoDataServlet")
 public class getPhotoDataServlet extends HttpServlet {
-    private static String api_key = "ms3MvC2UjwJBlSs5wNTVj-3SXPPAURq3";
-    private static  String api_secret = "P6wgCmBYbeFGRG76cwTOTe6k2V5jS1v";
-    private final static int CONNECT_TIME_OUT = 90000;
-    private final static int READ_OUT_TIME = 100000;
-    private static String boundaryString = getBoundary();
-
     PhotoService photoService;
     PhotoInfo photoInfo;
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
+        response.setHeader("content-type", "text/html;charset=UTF-8");
+        response.setCharacterEncoding("UTF-8");
         PrintWriter out = response.getWriter();
         BufferedReader reader = request.getReader();
+        JSONObject res = new JSONObject();
         String line = "";
         String content = "";
         while ((line = reader.readLine()) != null){
             content+=line;
         }
         String jbstring = content.toString();
-        System.out.println(jbstring);
+
         String temp[] = jbstring.split(",");
-        System.out.println(temp.length);
-//        System.out.println(temp.toString());
+
         String imageName = temp[0];
         String imgStr = temp[1];
-
-        System.out.println("hello"+imageName);
-
-        BASE64Decoder decoder = new BASE64Decoder();
-        try{
-            //Base64解码
-            byte[] b = decoder.decodeBuffer(imgStr);
-            for(int i = 0;i<b.length;i++){
-                if(b[i]<0){
-                    b[i]+=256;
+        String path = "D:\\Projects\\WebGIS\\GeoPic\\GeoPic\\web\\static\\data\\photos\\";
+        boolean judgeIsExistRes = judgeIsExist(imageName,path);
+        if(judgeIsExistRes){
+            res.put("message","图片已存在，不重复入库");
+            res.put("success",false);
+            out.write(res.toString());
+            return;
+        }else{
+            BASE64Decoder decoder = new BASE64Decoder();
+            try{
+                //Base64解码
+                byte[] b = decoder.decodeBuffer(imgStr);
+                for(int i = 0;i<b.length;i++){
+                    if(b[i]<0){
+                        b[i]+=256;
+                    }
                 }
+                String imagePath = "D:\\Projects\\WebGIS\\GeoPic\\GeoPic\\web\\static\\data\\photos\\"+imageName;
+                String thums = "D:\\Projects\\WebGIS\\GeoPic\\GeoPic\\web\\static\\data\\faces\\"+imageName;
+                System.out.println(imagePath);
+                OutputStream o = new FileOutputStream(imagePath);
+                o.write(b);
+                o.flush();
+                o.close();
+                res.put("message","可以入库");
+                res.put("success",true);
+                out.write(res.toString());
+            } catch (Exception e) {
+                e.printStackTrace();
             }
-            String imagePath = "D:\\Projects\\WebGIS\\GeoPic\\GeoPic\\web\\static\\data\\photos\\"+imageName;
-            String thums = "D:\\Projects\\WebGIS\\GeoPic\\GeoPic\\web\\static\\data\\faces\\"+imageName;
-            System.out.println(imagePath);
-            OutputStream o = new FileOutputStream(imagePath);
-            o.write(b);
-            o.flush();
-            o.close();
-//            BufferedImage bufferedImage =  ImageIO.read(new File(imagePath));
-//            BufferedImage subImage=bufferedImage.getSubimage(10,10,20,20);
-//            ImageIO.write(subImage,"JPEG",new File(thums));
-
-
-
-        } catch (Exception e) {
-            e.printStackTrace();
         }
-
-        out.write(jbstring);
 
     }
 
@@ -92,100 +89,21 @@ public class getPhotoDataServlet extends HttpServlet {
         doPost(request,response);
     }
 
-    protected static byte[] post(String url, HashMap<String, String> map, HashMap<String, byte[]> fileMap) throws Exception{
-        HttpURLConnection conne;
-        URL url1 = new URL(url);
-        conne = (HttpURLConnection) url1.openConnection();
-        conne.setDoOutput(true);
-        conne.setUseCaches(false);
-        conne.setRequestMethod("POST");
-        conne.setConnectTimeout(CONNECT_TIME_OUT);
-        conne.setReadTimeout(READ_OUT_TIME);
-        conne.setRequestProperty("accept", "*/*");
-        conne.setRequestProperty("Content-Type", "multipart/form-data; boundary=" + boundaryString);
-        conne.setRequestProperty("connection", "Keep-Alive");
-        conne.setRequestProperty("user-agent", "Mozilla/4.0 (compatible;MSIE 6.0;Windows NT 5.1;SV1)");
-        DataOutputStream obos = new DataOutputStream(conne.getOutputStream());
-        Iterator iter = map.entrySet().iterator();
-        while(iter.hasNext()){
-            Map.Entry<String, String> entry = (Map.Entry) iter.next();
-            String key = entry.getKey();
-            String value = entry.getValue();
-            obos.writeBytes("--" + boundaryString + "\r\n");
-            obos.writeBytes("Content-Disposition: form-data; name=\"" + key
-                    + "\"\r\n");
-            obos.writeBytes("\r\n");
-            obos.writeBytes(value + "\r\n");
-        }
-        if(fileMap != null && fileMap.size() > 0){
-            Iterator fileIter = fileMap.entrySet().iterator();
-            while(fileIter.hasNext()){
-                Map.Entry<String, byte[]> fileEntry = (Map.Entry<String, byte[]>) fileIter.next();
-                obos.writeBytes("--" + boundaryString + "\r\n");
-                obos.writeBytes("Content-Disposition: form-data; name=\"" + fileEntry.getKey()
-                        + "\"; filename=\"" + encode(" ") + "\"\r\n");
-                obos.writeBytes("\r\n");
-                obos.write(fileEntry.getValue());
-                obos.writeBytes("\r\n");
+    protected boolean judgeIsExist(String photoName,String photoDir){
+        boolean isExist = false;
+        File file = new File(photoDir);
+        File []array = file.listFiles();
+        for(int i = 0;i<array.length;i++){
+            if(array[i].isFile()){
+                String imgName = array[i].getName();
+                if(imgName.equals(photoName)){
+                    isExist = true;
+                    break;
+                }
             }
         }
-        obos.writeBytes("--" + boundaryString + "--" + "\r\n");
-        obos.writeBytes("\r\n");
-        obos.flush();
-        obos.close();
-        InputStream ins = null;
-        int code = conne.getResponseCode();
-        try{
-            if(code == 200){
-                ins = conne.getInputStream();
-            }else{
-                ins = conne.getErrorStream();
-            }
-        }catch (SSLException e){
-            e.printStackTrace();
-            return new byte[0];
-        }
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        byte[] buff = new byte[4096];
-        int len;
-        while((len = ins.read(buff)) != -1){
-            baos.write(buff, 0, len);
-        }
-        byte[] bytes = baos.toByteArray();
-        ins.close();
-        return bytes;
+        return isExist;
     }
 
-    private static String encode(String value) throws Exception{
-        return URLEncoder.encode(value, "UTF-8");
-    }
 
-    private static String getBoundary() {
-        StringBuilder sb = new StringBuilder();
-        Random random = new Random();
-        for(int i = 0; i < 32; ++i) {
-            sb.append("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_-".charAt(random.nextInt("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_".length())));
-        }
-        return sb.toString();
-    }
-
-    public static byte[] getBytesFromFile(File f) {
-        if (f == null) {
-            return null;
-        }
-        try {
-            FileInputStream stream = new FileInputStream(f);
-            ByteArrayOutputStream out = new ByteArrayOutputStream(1000);
-            byte[] b = new byte[1000];
-            int n;
-            while((n = stream.read(b)) != -1) {
-                out.write(b, 0, n);
-            }
-            stream.close();
-            out.close();
-            return out.toByteArray();
-        } catch (IOException e) {
-        }
-        return null;
-    }
 }
